requirements.txt
asgiref==3.11.0
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Pillow==12.0.0
PyJWT==2.10.1
sqlparse==0.5.4
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
whitenoise==6.8.2
dj-database-url==2.2.0
django-ratelimit==4.1.0

# Odoo Integration
odoorpc==0.10.1

# Celery + Redis (Async tasks + Cache)
redis==5.0.1
celery==5.3.6
django-redis==5.4.0
django-celery-beat==2.8.0
django-celery-results==2.6.0
flower==2.0.1
